refatoração da api
