+++ MessageQueueBasic.Api/Dockerfile
@@ -1,0 +1,22 @@
+#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
+
+FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
+WORKDIR /app
+EXPOSE 80
+EXPOSE 443
+
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
+WORKDIR /src
+COPY ["MessageQueueBasic.Api/MessageQueueBasic.Api.csproj", "MessageQueueBasic.Api/"]
+RUN dotnet restore "MessageQueueBasic.Api/MessageQueueBasic.Api.csproj"
+COPY . .
+WORKDIR "/src/MessageQueueBasic.Api"
+RUN dotnet build "MessageQueueBasic.Api.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "MessageQueueBasic.Api.csproj" -c Release -o /app/publish
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "MessageQueueBasic.Api.dll"]
