+++ MessageQueueBasic.Api/MessageQueueBasic.Api.csproj
