+++ MessageQueueBasic.Domain/MessageQueueBasic.Domain.csproj
@@ -7,4 +7,8 @@
     <LangVersion>preview</LangVersion>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\" />
+  </ItemGroup>
+
 </Project>
