+++ MessageQueueBasic.MessageBroker/Consumers/TestConsumer.cs
@@ -4,17 +4,12 @@ using Microsoft.Extensions.Logging;
 
 namespace MessageQueueBasic.MessageBroker.Consumers;
 
-public class TestConsumer : IConsumer<ITested>
+public class TestConsumer(ILogger<TestConsumer> logger) : IConsumer<ITested>
 {
-    readonly ILogger<TestConsumer> _logger;
-
-    public TestConsumer(ILogger<TestConsumer> logger)
-    {
-        _logger = logger;
-    }
     public async Task Consume(ConsumeContext<ITested> context)
     {
-        _logger.LogInformation("Mensagem recebida", context.Message.Name);
+        string message = $"Mensagem recebida {context.Message.Name}";
+        logger.LogInformation(message);
         await Task.FromResult(context.Message);
     }
 }
