--- MessageQueueBasic.Web/Dockerfile
@@ -1,22 +1,0 @@
-#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
-WORKDIR /app
-EXPOSE 80
-EXPOSE 443
-
-FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
-WORKDIR /src
-COPY ["MessageQueueBasic.Web/MessageQueueBasic.Web.csproj", "MessageQueueBasic.Web/"]
-RUN dotnet restore "MessageQueueBasic.Web/MessageQueueBasic.Web.csproj"
-COPY . .
-WORKDIR "/src/MessageQueueBasic.Web"
-RUN dotnet build "MessageQueueBasic.Web.csproj" -c Release -o /app/build
-
-FROM build AS publish
-RUN dotnet publish "MessageQueueBasic.Web.csproj" -c Release -o /app/publish
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "MessageQueueBasic.Web.dll"]
