+++ MessageQueueBasicSettings.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32630.194
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34322.80
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageQueueBasic.Domain", "MessageQueueBasic.Domain\MessageQueueBasic.Domain.csproj", "{D491BAA8-B1F6-4EE7-B552-A965702903EC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageQueueBasic.MessageBroker", "MessageQueueBasic.MessageBroker\MessageQueueBasic.MessageBroker.csproj", "{EC46B75E-6E46-4004-92DB-93B4E4DA3AF8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageQueueBasic.Web", "MessageQueueBasic.Web\MessageQueueBasic.Web.csproj", "{2B4E24D5-DE8A-4753-96ED-1D0D9F68071C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageQueueBasic.Api", "MessageQueueBasic.Api\MessageQueueBasic.Api.csproj", "{9F8E83EB-8751-4D09-B225-E9EF1AC4F19A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{EC46B75E-6E46-4004-92DB-93B4E4DA3AF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EC46B75E-6E46-4004-92DB-93B4E4DA3AF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EC46B75E-6E46-4004-92DB-93B4E4DA3AF8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2B4E24D5-DE8A-4753-96ED-1D0D9F68071C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B4E24D5-DE8A-4753-96ED-1D0D9F68071C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B4E24D5-DE8A-4753-96ED-1D0D9F68071C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B4E24D5-DE8A-4753-96ED-1D0D9F68071C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F8E83EB-8751-4D09-B225-E9EF1AC4F19A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F8E83EB-8751-4D09-B225-E9EF1AC4F19A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F8E83EB-8751-4D09-B225-E9EF1AC4F19A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F8E83EB-8751-4D09-B225-E9EF1AC4F19A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
